V 0.1 initial commit
Minimally functional web scraper that extracts all of the cases cited in a
reported decision on CanLII.

Accepts a URL from a CanLII decision as input and returns all of the cases that
 the decision cites. Current version uses the CanLII API to pull and display a
JSON file from the service.

The current version is limited to "casesCited" calls. The casesCited call does
not pick up cases that are not indexed on CanLII, so several results will be
missing