Version 0.3 update
Updated with a wholly new modular layout. The main data_miner.py
program has been replaced by standard_report_generator.py and the
accompanying module libraries in the new apps/ directory.

standard_report_generator.py will eventually produce a report that
is compatible with the appeal court research project, at which
point this part of the amicus project will be where it needs to be